requirements.txt
Django==5.0.1
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
gunicorn==22.0.0